Added menu, resources, easygui.py
Added a simple menu. Used easygui.py to impliment it so it is needed on
your machines. There is a slight issue with an infinite loop when the
menu is ran in main....still trying to figure that shit out. The sounds
seem to be loading correctly under defINIT() but just in case I added
them outside of it and commented them out for now. Please let me know if
you have a better idea or some input on what you want to see with the
menu.