Bringing in changes
I was tired and didn't bother with the github with this. I uploaded a
new working GUI on the class file exchange site. I will be finishing the
AI code this week and then finishing up the single player mode. Someone
please needs to get to work on the netcode stuff, I don't have time to
spend doing that and finishing up single player. Ryan you said doing
netcode would be easy with python so maybe you can work on that. In case
you guys don't check the file exchange I included the 7z file. Oh and we
need to get our testsuite done and updated tasklist or something for
next tuesday.