Ok fixed the AI bugs I think.
PLEASE PLAYTEST THIS, I'm like 99% sure the new code is good. It is long
and a pain in the butt, but seems to work fine. I played with it for
about 30 minutes without issue, but who knows. Let me know if you find
bugs, or if you get ambitious try to error check it yourself.